+++ onnx/defs/schema.h
@@ -18,9 +18,7 @@
 
 #include "data_type_utils.h"
 #include "onnx/checker.h"
-#include "onnx/common/status.h"
 #include "onnx/defs/shape_inference.h"
-#include "onnx/onnx-operators_pb.h"
 
 namespace ONNX_NAMESPACE {
 using namespace checker;
@@ -668,40 +666,5 @@ class OpSchemaRegistry final {
 
 // Helper function
 size_t ReplaceAll(std::string& s, const char* from, const char* to);
-
-typedef Common::Status (*BuildFunction)(std::shared_ptr<FunctionProto>*);
-
-class FunctionBuilder {
- public:
-  FunctionBuilder& SetDomain(const std::string& domain);
-  const std::string& GetDomain() const;
-  FunctionBuilder& SetBuildFunction(BuildFunction build_func);
-
-  BuildFunction GetBuildFunction() const;
-
- private:
-  std::string domain_;
-  BuildFunction build_func_;
-};
-
-class FunctionBuilderRegistry {
- public:
-  FunctionBuilderRegistry() = default;
-
-  void Register(const FunctionBuilder& function_builder);
-
-  // Get functions for specific domain.
-  Status GetFunctions(
-      const std::string& domain,
-      /*out*/
-      std::multimap<std::string, std::shared_ptr<FunctionProto>>* function_set)
-      const;
-
-  static FunctionBuilderRegistry& OnnxInstance();
-
- private:
-  std::vector<FunctionBuilder> function_builders;
-  std::mutex mutex_;
-}; // namespace ONNX_NAMESPACE
 
 } // namespace ONNX_NAMESPACE
