+++ onnx/onnx-operators_pb.h
@@ -1,13 +1,11 @@
 // Copyright (c) Facebook Inc. and Microsoft Corporation.
 // Licensed under the MIT license.
 
-#ifndef ONNX_ONNX_OPERATOR_PB_H
-#define ONNX_ONNX_OPERATOR_PB_H
+#pragma once
 
+#include "onnx/onnx_pb.h"
 #ifdef ONNX_ML
 #include "onnx/onnx-operators-ml.pb.h"
 #else
 #include "onnx/onnx-operators.pb.h"
 #endif
-
-#endif // ! ONNX_ONNX_OPERATOR_PB_H
